CID-2058: Reseting pin failure count on pin update
- Resets pin failure count when pin is reset or updated
- Setting the pin to the same value does not reset the counter
- Can not change the pin for a user whose PIN is locked

+++ src/nonCommitTest/groovy/com/rackspace/idm/api/resource/cloud/v20/ResetPhonePinForUserIntegrationTest.groovy
@@ -1,8 +1,16 @@
 package com.rackspace.idm.api.resource.cloud.v20
 
+import com.rackspace.docs.identity.api.ext.rax_auth.v1.PhonePinStateEnum
 import com.rackspace.idm.Constants
+import com.rackspace.idm.ErrorCodes
+import com.rackspace.idm.GlobalConstants
 import com.rackspace.idm.domain.config.IdentityConfig
+import com.rackspace.idm.domain.entity.User
+import com.rackspace.idm.domain.service.IdentityUserService
+import org.openstack.docs.identity.api.v2.ForbiddenFault
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.test.context.ContextConfiguration
+import testHelpers.IdmAssert
 import testHelpers.RootIntegrationTest
 
 import javax.ws.rs.core.MediaType
@@ -11,6 +19,9 @@ import static javax.servlet.http.HttpServletResponse.*
 
 @ContextConfiguration(locations = "classpath:app-config.xml")
 class ResetPhonePinForUserIntegrationTest extends RootIntegrationTest {
+
+    @Autowired
+    IdentityUserService identityUserService
 
     def "reset phone pin"() {
         given:
@@ -181,6 +192,60 @@ class ResetPhonePinForUserIntegrationTest extends RootIntegrationTest {
 
         and: "phone pin is not changed"
         phonePin == updatedPhonePin
+
+        cleanup:
+        utils.deleteUser(user)
+
+        where:
+        contentType << [MediaType.APPLICATION_XML_TYPE, MediaType.APPLICATION_JSON_TYPE]
+    }
+
+    def "resetPhonePin: reset phone pin updates invalid count" () {
+        given:
+        def user = utils.createGenericUserAdmin()
+
+        // Update user to have failed pin attempts
+        User userEntity = identityUserService.getEndUserById(user.id)
+        userEntity.recordFailedPinAuthentication()
+        identityUserService.updateEndUser(userEntity)
+        assert userEntity.phonePinAuthenticationFailureCount > 0
+
+        when: "reset phone pin"
+        def response = cloud20.resetPhonePin(utils.getServiceAdminToken(), user.id, false, contentType)
+
+        then: "failure count is reset"
+        assert response.status == SC_NO_CONTENT
+
+        and:
+        User userEntityAfter = identityUserService.getEndUserById(user.id)
+        userEntityAfter.phonePinAuthenticationFailureCount == 0
+
+        cleanup:
+        utils.deleteUser(user)
+
+        where:
+        contentType << [MediaType.APPLICATION_XML_TYPE, MediaType.APPLICATION_JSON_TYPE]
+    }
+
+    def "resetPhonePin: can not reset phone pin on locked account" () {
+        given:
+        def user = utils.createGenericUserAdmin()
+        def token = utils.getToken(user.username, Constants.DEFAULT_PASSWORD)
+
+        // Update user to have failed pin attempts
+        User userEntity = identityUserService.getEndUserById(user.id)
+
+        GlobalConstants.PHONE_PIN_AUTHENTICATION_FAILURE_LOCKING_THRESHOLD.times {
+            userEntity.recordFailedPinAuthentication()
+        }
+        identityUserService.updateEndUser(userEntity)
+        assert userEntity.getPhonePinState() == PhonePinStateEnum.LOCKED
+
+        when: "reset phone pin"
+        def response = cloud20.resetPhonePin(utils.getServiceAdminToken(), user.id, false, contentType)
+
+        then: "Not allowed"
+        IdmAssert.assertOpenStackV2FaultResponse(response, ForbiddenFault, SC_FORBIDDEN, ErrorCodes.ERROR_CODE_PHONE_PIN_LOCKED, ErrorCodes.ERROR_MESSAGE_PHONE_PIN_LOCKED)
 
         cleanup:
         utils.deleteUser(user)
